Now compatible with esx_advancedgarage
Edited code to look for plate when removing and moving vehicle from owned_vehicles to impounded_vehicles. Before it was looking for id column but with the way the vehicle shop it is set. It was creating mysql error and would not remove vehicle.  Anyway, now it will work properly. Also edited code to look for stored instead of state as i use esx_advancegarage and not esx_eden_garage. So i do not have any state column.